Restructure TUI into separate files with bug fixes, UX improvements and tests
Split 651-line main.go into 6 focused files (main.go, model.go, update.go,
view.go, commands.go, styles.go). Fix PREFIX trailing slash normalisation,
remove dead value-receiver mutations, delete custom max() in favour of
builtin, replace ListObjectsV2Paginator with direct API call.

Add auto-skip bucket selection for single SES bucket, Esc back-navigation,
empty inbox state, email count indicator, consistent date formatting,
vertical email header, r to refresh, delete status feedback. Replace
isDeleteFromList with previousState. Extract renderEmailView helper.

29 unit tests covering cutToWidth, placeOverlay, prefix normalisation,
renderListHelp, bucket auto-selection, email deletion, sorting, and
status message lifecycle.
